requirements.txt
fastapi[standard]==0.104.1
uvicorn==0.24.0
assemblyai==0.17.0
python-dotenv==1.0.0
pydantic==2.5.1
librosa
numpy
soundfile
